requirements.txt
# LinguaAI Backend Dependencies

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Firebase
firebase-admin==6.4.0

# AI/ML - Speech to Text
openai-whisper==20231117

# AI/ML - Text to Speech (gTTS - Google Text-to-Speech)
gTTS>=2.5.0

# Phoneme processing
phonemizer==3.2.1

# Text similarity
python-Levenshtein==0.23.0

# Utilities
numpy==1.26.3
tqdm==4.66.1
